Use FullMarket from common/, refactor
