+++ discord-bot/tsconfig.json
@@ -9,13 +9,22 @@
         "moduleResolution": "node",
         "outDir": "./lib",
         "tsBuildInfoFile": "lib/tsconfig.tsbuildinfo",
-        "baseUrl": "./src",
         "strict": true,
-    },
+        "paths": {
+            "common/*": ["../common/src/*", "../../common/lib/*"],
+            "discord-bot/*": ["./src/*"]
+        }
+         },
     "ts-node": {
         "require": ["tsconfig-paths/register"],
         "esm": true
     },
+    "references": [
+        {
+            "path": "../common"
+        },
+    ],
+
     "compileOnSave": true,
     "include": ["src/**/*.ts"]
 }
